requirements.txt
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
crcmod==1.7
cryptography==2.6.1
decorator==5.0.7
gitdb==4.0.7
google-trans-new==1.1.9
googletrans==3.0.0
httpx==0.13.3
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
jsonschema==3.2.0
joblib==1.0.1
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.1
MultiApp==1.0
pdf2image==1.14.0
pdfminer==20191125
pdfminer.six==20200517
pdfplumber==0.5.28
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
Pygments==2.9.0
pyPdf==1.13
PyPDF2==1.26.0
pyrsistent==0.17.3
pytesseract==0.3.7
Flask==1.1.2
gitdb==4.0.7
GitPython==3.1.15
idna==2.10
numpy==1.20.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
sklearn
matplotlib>=1.4.3
pandas>=0.19
flasgger==0.9.4
streamlit
